Update Termux-ADB-SSH-StartUp_Connection-Info-Widget_Stars_version.sh
Displays service stopped or the IP port info depending on if the ports were found open in the background by grepping.

Widget only displays the status. The start and stopping is controlled by different script in termux shortcut launching widget or Termux Tasker launcher from the shutdown reboot screen. 

All starting and stopping script should be in .shortcuts in termux.